requirements/dev-requirements.txt
autopep8==1.4.3
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pycodestyle==2.5.0
PyJWT==1.7.1
SQLAlchemy==1.3.1
Werkzeug==0.14.1
certifi==2019.3.9 
chardet==3.0.4 
idna==2.8 
requests==2.21.0 
urllib3==1.24.1
